Adding leaflet map

See merge request scrozprosjecni/MyVinylCollection!13

+++ frontend/src/pages/ProfilePage.jsx
@@ -7,6 +7,7 @@ import EditIcon from "@mui/icons-material/Edit";
 
 import {getCurrentUser, IsMobile} from "../util/utils";
 import authHeader from "../auth-header";
+import LocationMap from "../components/LocationMap";
 
 const infoContainerStyle = {
   display: "flex",
@@ -38,12 +39,16 @@ function ProfilePage() {
 
   const [name, setName] = React.useState();
   const [surname, setSurname] = React.useState();
-  const [location, setLocation] = React.useState();
-  const [country, setCountry] = React.useState();
-  const [city, setCity] = React.useState();
+  const [location, setLocation] = React.useState({});
+  const [latitude, setLatitude] = React.useState();
+  const [longitude, setLongitude] = React.useState();
   const [email, setEmail] = React.useState();
   const [contactEmail, setContactEmail] = React.useState();
+
   const [password, setPassword] = React.useState();
+  const [newPassword, setNewPassword] = React.useState();
+  const [confirmPassword, setConfirmPassword] = React.useState();
+  const [errorMessage, setErrorMessage] = React.useState()
 
   React.useEffect(() => {
     fetch(api + `/users/info/${username}`, {
@@ -53,7 +58,13 @@ function ProfilePage() {
         Authorization: authHeader(),
       },
     }).then((response) => {
-      response.json().then((result) => setData(result));
+      response.json().then((result) => {
+        setData(result)
+        console.log(result)
+        setLatitude(result.location.latitude)
+        setLongitude(result.location.longitude)
+        console.log(result.location.latitude + "  " + result.location.longitude)
+      });
     });
   }, [api, username]);
 
@@ -65,8 +76,12 @@ function ProfilePage() {
       setEmail(data.email);
       setContactEmail(data.contactEmail);
       setPassword(data.password);
-      setTimeout(() => setLoading(false), 500);
+      console.log(data.password)
+      setTimeout(() => {
+        setLoading(false)
+      }, 500);
     } else {
+      console.log(data)
       localStorage.setItem("username", username);
       fetch(api + `/users/info/${username}`, {
         method: "PUT",
@@ -80,52 +95,28 @@ function ProfilePage() {
     }
   }, [data]);
 
-  React.useEffect(() => {
-    if (data.location !== undefined) {
-      setCountry(data.location.substr(0, data.location.indexOf(",")));
-      setCity(
-        data.location.substr(
-          data.location.indexOf(",") + 2,
-          data.location.length
-        )
-      );
-    }
-  }, [data.location]);
-
-  const [countryData, setCountryData] = React.useState({});
-  let c = [];
-  const [countries, setCountries] = React.useState([]);
-  const [cities, setCities] = React.useState([]);
-
-  React.useEffect(() => {
-    fetch("https://countriesnow.space/api/v0.1/countries", {
-      method: "GET",
-    })
-      .then(async (response) => {
-        setCountryData(await response.json());
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-  }, []);
-
-  React.useEffect(() => {
-    if (countryData.data !== undefined)
-      countryData.data.forEach((d) => c.push(d.country));
-    setCountries(c);
-  }, [countryData]);
-
   function saveChanges() {
-    setEditingMode(false);
-    localStorage.setItem("username", username);
-    setData({
-      name: name,
-      surname: surname,
-      username: username,
-      location: `${country}, ${city}`,
-      email: email,
-      password: password,
-    });
+    if (confirmPassword === newPassword) {
+      setEditingMode(false);
+      localStorage.setItem("username", username);
+      setErrorMessage("")
+      setConfirmPassword("")
+      setData({
+        name: name,
+        surname: surname,
+        username: username,
+        location: {
+          latitude: latitude,
+          longitude: longitude
+        },
+        email: email,
+        password: newPassword,
+      });
+    } else {
+      setErrorMessage("Passwords must match");
+      setConfirmPassword("")
+      setPassword("")
+    }
   }
 
   function cancelChanges() {
@@ -138,22 +129,12 @@ function ProfilePage() {
     setEditingMode(false);
   }
 
-  function handleCountryChange(value) {
-    setCities([]);
-    setCountry(value);
-    setCity("");
-    setCities(
-      countryData.data.filter((data) => {
-        return data.country === value;
-      })[0].cities
-    );
-  }
-
   return (
-    <div
+    <form
       style={
         mobile ? { ...containerStyle, marginBottom: "4rem" } : containerStyle
       }
+      autocomplete="off"
     >
       <ProfileHeader />
       {loading ? (
@@ -208,34 +189,31 @@ function ProfilePage() {
             value={contactEmail}
             onChange={(e) => setEmail(e.target.value)}
           />
-          <TextField
-            style={{ marginTop: "2rem" }}
-            label="Password"
-            type="password"
-            disabled={!editingMode}
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-          <Autocomplete
-            style={{ marginTop: "2rem" }}
-            value={country}
-            disabled={!editingMode}
-            renderInput={(params) => (
-              <TextField {...params} autoComplete="off" label="Country" />
-            )}
-            options={countries}
-            onChange={(event, value) => handleCountryChange(value)}
+
+          <LocationMap
+              lat={data.location.latitude} lng={data.location.longitude}
+              setLat={setLatitude} setLng={setLongitude}
+              editing={editingMode}
           />
 
-          <Autocomplete
-            style={{ marginTop: "2rem" }}
-            value={city}
-            disabled={!editingMode}
-            renderInput={(params) => (
-              <TextField {...params} autoComplete="off" label="City" />
-            )}
-            options={cities}
-            onChange={(event, value) => setCity(value)}
+          <p style={{color: "red", fontSize: "12px", marginTop: "1rem"}}>{errorMessage}</p>
+
+          <TextField
+              label="New Password"
+              type="password"
+              disabled={!editingMode}
+              autoComplete={false}
+              value=""
+              onChange={(e) => setNewPassword(e.target.value)}
+          />
+
+          <TextField
+              style={{ marginTop: "2rem" }}
+              label="Confirm Password"
+              type="password"
+              disabled={!editingMode}
+              autoComplete={false}
+              onChange={(e) => setConfirmPassword(e.target.value)}
           />
 
           <div style={{ marginTop: "1rem" }}>
@@ -260,7 +238,7 @@ function ProfilePage() {
           </div>
         </div>
       )}
-    </div>
+    </form>
   );
 }
